third_party/blink/web_tests/external/wpt/css/css-values/minmax-time-serialize-expected.txt

This is a testharness.js-based test.
FAIL 'min(1ms)' as a specified value should serialize as 'calc(0.001s)'. assert_equals: 'min(1ms)' and 'calc(0.001s)' should serialize the same in specified values. expected "calc(0.001s)" but got "min(1ms)"
PASS 'min(1ms)' as a computed value should serialize as '0.001s'.
FAIL 'min(1s)' as a specified value should serialize as 'calc(1s)'. assert_equals: 'min(1s)' and 'calc(1s)' should serialize the same in specified values. expected "calc(1s)" but got "min(1s)"
PASS 'min(1s)' as a computed value should serialize as '1s'.
FAIL 'max(1ms)' as a specified value should serialize as 'calc(0.001s)'. assert_equals: 'max(1ms)' and 'calc(0.001s)' should serialize the same in specified values. expected "calc(0.001s)" but got "max(1ms)"
PASS 'max(1ms)' as a computed value should serialize as '0.001s'.
FAIL 'max(1s)' as a specified value should serialize as 'calc(1s)'. assert_equals: 'max(1s)' and 'calc(1s)' should serialize the same in specified values. expected "calc(1s)" but got "max(1s)"
PASS 'max(1s)' as a computed value should serialize as '1s'.
FAIL 'min(1s, 2s, 3s)' as a specified value should serialize as 'calc(1s)'. assert_equals: 'min(1s, 2s, 3s)' and 'calc(1s)' should serialize the same in specified values. expected "calc(1s)" but got "min(1s, 2s, 3s)"
PASS 'min(1s, 2s, 3s)' as a computed value should serialize as '1s'.
FAIL 'min(3s, 2s, 1s)' as a specified value should serialize as 'calc(1s)'. assert_equals: 'min(3s, 2s, 1s)' and 'calc(1s)' should serialize the same in specified values. expected "calc(1s)" but got "min(3s, 2s, 1s)"
PASS 'min(3s, 2s, 1s)' as a computed value should serialize as '1s'.
FAIL 'max(1s, 2s, 3s)' as a specified value should serialize as 'calc(3s)'. assert_equals: 'max(1s, 2s, 3s)' and 'calc(3s)' should serialize the same in specified values. expected "calc(3s)" but got "max(1s, 2s, 3s)"
PASS 'max(1s, 2s, 3s)' as a computed value should serialize as '3s'.
FAIL 'max(3s, 2s, 1s)' as a specified value should serialize as 'calc(3s)'. assert_equals: 'max(3s, 2s, 1s)' and 'calc(3s)' should serialize the same in specified values. expected "calc(3s)" but got "max(3s, 2s, 1s)"
PASS 'max(3s, 2s, 1s)' as a computed value should serialize as '3s'.
FAIL 'min(900ms, 1s)' as a specified value should serialize as 'calc(0.9s)'. assert_equals: 'min(900ms, 1s)' and 'calc(0.9s)' should serialize the same in specified values. expected "calc(0.9s)" but got "min(900ms, 1s)"
PASS 'min(900ms, 1s)' as a computed value should serialize as '0.9s'.
FAIL 'max(1100ms, 1s)' as a specified value should serialize as 'calc(1.1s)'. assert_equals: 'max(1100ms, 1s)' and 'calc(1.1s)' should serialize the same in specified values. expected "calc(1.1s)" but got "max(1100ms, 1s)"
PASS 'max(1100ms, 1s)' as a computed value should serialize as '1.1s'.
FAIL 'calc(min(1s, 2s) + max(3s, 4s) + 10s)' as a specified value should serialize as 'calc(15s)'. assert_equals: 'calc(min(1s, 2s) + max(3s, 4s) + 10s)' and 'calc(15s)' should serialize the same in specified values. expected "calc(15s)" but got "calc(10005ms)"
FAIL 'calc(min(1s, 2s) + max(3s, 4s) + 10s)' as a computed value should serialize as '15s'. assert_equals: 'calc(min(1s, 2s) + max(3s, 4s) + 10s)' and '15s' should serialize the same in computed values. expected "15s" but got "10.005s"
Harness: the test ran to completion.
